Fix querying for events and issues, part 1 of N
The query url for events and issues lived under `/api/0/organizations`:

   /api/0/organizations/$ORG_SLUG/events/?$QUERY_PARAMS
   /api/0/organizations/$ORG_SLUG/issues/?$QUERY_PARAMS

These endpoint don't return results anymore, and don't exist in the
Sentry API documentation.

New endpoints that do return resutls exist for events and issues:

   /api/0/projects/$ORG_SLUG/$PROJECT_ID/events/?$QUERY_PARAMS
   /api/0/projects/$ORG_SLUG/$PROJECT_ID/issues/?$QUERY_PARAMS

This PR updates queries to use the new API endpoints.

Also, event decoding expected an object with a key `data`, and a list of
event objects as the value. The new API endpoint returns a list of events,
so this PR updates the decoding to expect that.

This PR is enough to make queries run, and render them, but it's not
enough. Followup PRs are required to fix the interface, because it
currently assumes an organization context, with a list of projects.