Deploy to Grafana Cloud only from publish workflow
This PR adds `scopes: grafana_cloud` to our publish workflow. The build
artifacts will only be available in Grafana Cloud, enabling us to test
the release there before releasing for OSS and on-prem.

The release docs here have been updated to include a section for Sentry
when this PR is merged. tl;dr: an extra command will need to be run to
publish a version to on-prem and OSS:

    $ SENTRY_VERSION=2.2.1
    $ gcom /plugins/grafana-sentry-datasource/versions/$SENTRY_VERSION -XPOST -d scopes=universal